irqchip: Add driver for Cirrus Logic Tacna codecs
The Cirrus Logic Tacna codecs are highly complex devices containing
programmable DSPs and many other internal sources of interrupts plus a
number of GPIOs that can be used as interrupt inputs. The large number
(>150) of interrupt sources are managed by an on-board interrupt
controller.

This driver provides the handling for the interrupt controller. As the
codec is accessed via regmap, we can make use of the generic IRQ
functionality from regmap to do most of the work.

The KConfig options are not user-configurable because this driver is
mandatory so is automatically included when the parent MFD driver is
selected.

Change-Id: Ib71c574027dcb9af61428d5db57c203ef8832d63
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>