ASoC: cs47l94: Basic support for the ultrasonic demodulators
Add basic support for the ultrasonic demodulators. They are a standard
mixer input, but only take IN1-4 L/R as inputs. They can operate
independently of whether the associated mixer input is part of a standard
audio path. Note that the selected bandwidth will alter the output
frequency as follows:
0 = 24.5kHz-40.5kHz : Fout=32kHz
1 = 18kHz-22kHz : Fout=8kHz
2 = 16kHz-24kHz : Fout=16kHz
3 = 20kHz-28kHz : Fout=16kHz

Change-Id: I5f23ee0a373e6d6059284f12a0c09864d1641256
Signed-off-by: Stuart Henderson <stuarth@opensource.wolfsonmicro.com>
Signed-off-by: Swetha Nagapuri <swetha.nagapuri@soctronics.com>