ASoC: tacna: Make mono output routing codec-specific
When an output is selected for bridged mono the widgets that
must be routed together depends on the output configuration
of the codec. Remove the simple default table and instead have
the codec driver supply a table of routing and allow this to
supply multiple routes for each output so that multiple widgets
can be routed together if needed to handle the outputs on that
codec.

On outputs with a demux the initial mono bit setting should be
for the default demux position only, so that it is correct if
the demux control is left at default. If the demux control is
changed it will handle the mono bit state. Setting cfg_reg = 0
in struct tacna_mono_route skips initializing that output from
the corresponding pdata entry.

Change-Id: Ibb9357205cfba4fe6f3cc1bab72a06668661c766
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>