ASoC: tacna: DAC should depend on ASYNCCLK when in async domain
If any of the DACs are on the async domain they depend on ASYNCCLK
so there should be a DAPM route to the ASYNCCLK widget. But if they
are not on the ASYNCCLK domain we don't want to enable ASYNCCLK:
it wastes power and it might not have a source clock if it's not
intended to be used.

It isn't possible for a supply widget to go through a conditional
route, and it's difficult to find a decoy route that will power-up
conditionally with the DAC without creating side effects or lots
of additional widgets. Instead a dummy DAC_ASYNCCLK supply widget
has been added and both ASYNCCLK widgets use a custom handler to
manage the single enable bit. ASYNCCLK is enabled if either:

1) the ASYNCCLK widget is on, or
2) the DAC_ASYNCCLK widget is on and one or both of the DAC rate
   controls are on ASYNC.

If the DAC is being switched to an ASYNC domain the ASYNCCLK must
be started before switching domain. If it is being switched to a
SYNC domain the ASYNCCLK must be stopped after switching domain.

Change-Id: Ibecac2e3d43310ed545b9f36bcac2c89f3d04b75
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>