ASoC: clsic-codec: Remove unnecessary storage of DSP power tables
The clsic_dsp2_sram_power_regs table was stored in a member of
struct tacna_priv but the clsic-codec never actually uses the
tacna memory power functions. Instead it has a different
custom power-up sequence which is in the clsic-codec file so
can refer to the clsic_dsp2_sram_power_regs table directly.

Change-Id: I4b8e319cdb423849411497c5869f6737ff2d1e4c
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>