mfd: tacna: Don't get VDD_D regulator exclusive
An exclusive get of VDD_D was intended to prevent the
regulator being shared with other peripherals, which would
mean we couldn't be sure if the regulator would turn off in
suspend. But it's perfectly valid for the same regulator to
supply multiple supply pins on our codec, which would fail
the exclusive get.

Change-Id: I793d1edb9263b38275430b098f401cbd084df742
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>