mfd: clsic: remove unreachable debugcontrol code from worker thread
The code suggests that when the worker thread runs that the overall
driver state would be debugcontrol requested; however this could not be
the case as the pm_resume call wil have set state to resuming. Also, the
function that attempts to activate debugcontrol waits for the state to
become ON before continuing.

Change-Id: I3840bb4599d93f44df96e6ff7f27a5cbfcb16cd8
Signed-off-by: Simon Trimmer <simont@opensource.cirrus.com>