mfd: clsic: changes to hw and sw reset delays
The half second reset delay time originated from the FPGA and is clearly
too long for both hardware and software reset use on silicon.

Hardware reset has a short delay built into the disable_hard_reset()
call, put a similar short in the hard_reset() function in between the
enable and disable reset line to make sure it settles.

Similarly, remove the delay between applying the software reset and
polling for boot done as the polling will handle the delay.

Change-Id: I9594aea86ae953dee60b39b9e6364eeff927401c
Signed-off-by: Simon Trimmer <simont@opensource.cirrus.com>