mfd: tacna: Get all dependencies before registering children
If we EPROBE_DEFER after registering any child devices the children
will be removed and this causes the kernel to retry deferred drivers.
This can cause a loop where every defer immediately causes another
probe.

To avoid this, get the reset gpio and regulators early so that if
they aren't available we can defer before we have created any child
drivers.

Change-Id: I86d84f8f52392c3f52547be7d8822e3bfd2f2344
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>