ASoC: clsic: refactor setting of error info to be more logical
Using the separate function for snd_ctl_notify, make sure that error info
is set in appropriate places and that the notification only reaches user space
after the error info has been set correctly. This ensures the start and finish
of any driver interaction with user space is started with setting the driver
state and finished with a notification regarding the error info.

Change-Id: I74ab36f1458ca03da41caab75c8ba2b4c048395d
Signed-off-by: Ralph Clark <ralph.clark@cirrus.com>