ASoC: clsic-vox: simplify vox_notification_handler()
The code for the trigger detect notification was intermingled; moved the
comment to the top as it's describing the clsic_mode has already changed
in the device and there needed to be an explanation as to why
clsic_msgproc_use() was being called at that point.

Repetition complete is basically reporting success or failure; we didn't
need to verify the response codes coming from the device as the
userspace code is better placed to interpret and decide what to do with
them.

Change-Id: I5d37d8a6eba9a593906556684d16a6288e7d2bc3
Signed-off-by: Simon Trimmer <simont@opensource.cirrus.com>