mfd: clsic: enable programming blank flashes
When using an erased flash chip switching on the device will cause a boot
failure notification, this will switch the driver state into HALTED and prevent
further access.

However, the bootloader needs to progress programming the flash so allow
the bootloader to operate in the HALTED state and set the state to
RESUMING once the mab has been successfully downloaded to allow the
maintenance thread to do what is necessary.

Change-Id: I345d602c71c2799320e77b75c6c0113a52ea17d8
Signed-off-by: Simon Trimmer <simont@opensource.cirrus.com>