ASoC: clsic-vox: Fix memory duplication of platform data
Whenever the mfd_add_devices API is called the platform
data provided is duplicated causing mfd children to use
different platform data memory to that of their parents.
This leads to the clsic data structure becoming out of sync.

To avoid this a new API has been added allowing the child
to obtain a pointer to it's associated service handler
struct using the service instance as a unique identifier.

Change-Id: I5c85391a2ab82fef6a8e77d140654d49cbbe37fd
Signed-off-by: Lucas Tanure <tanureal@opensource.cirrus.com>