mfd: clsic: Don't block probe waiting for firmware download
The firmware download could take a while, it can also depend on
other parts of the system that are not yet available. Waiting for
completion can result in probe getting blocked for an extended
period, and this is simply wasting time if the system is not in
a state where the download could be done successfully anyway.

Change the call from probe to be asynchronous.

Change-Id: I31d57bdd2cfa9232ada271cae4b5977e5125291c
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>